requirements-rt2.2.txt
tensorflow==2.3.1
numpy==1.18.5
pandas==1.2.1
scikit-learn==0.23.1
google-api-python-client==1.9.3
gcsfs==0.7.0
google-cloud==0.34.0
google-cloud-storage==1.29.0
matplotlib==3.2.1
cloudml-hypertune
